pyproject.toml
[project]
name = "poc-llm-chat-and-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "jinja2>=3.1.6",
    "pydantic-ai-slim>=1.59.0",
    "pydantic-settings>=2.13.0",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
mistral = [
    "pydantic-ai-slim[mistral]>=1.59.0",
]
openai = [
    "pydantic-ai-slim[openai]>=1.59.0",
]
web = [
    "pydantic-ai-slim[web]>=1.59.0",
]
